requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.6.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
limits==5.1.0
mccabe==0.7.0
mongoengine==0.29.1
multidict==6.4.3
packaging==25.0
passlib==1.7.4
platformdirs==4.3.7
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.12.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.3
requests==2.32.3
resend==2.8.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tomli==2.2.1
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.2
websockets==14.2
wrapt==1.17.2
yarl==1.20.0
